Move reversescan tests to test_scans
Group tests with the other test for the
scan.py module.

JIRA: MKT-130, MT-2555
